+++ pkg/deployment/reconcile/plan_builder_rotate_upgrade.go
@@ -23,17 +23,9 @@
 package reconcile
 
 import (
-	"fmt"
-	"reflect"
-	"strings"
-
-	"k8s.io/apimachinery/pkg/api/equality"
-
-	"github.com/arangodb/kube-arangodb/pkg/apis/deployment"
-	"github.com/arangodb/kube-arangodb/pkg/deployment/pod"
-
 	"github.com/arangodb/go-driver"
 	upgraderules "github.com/arangodb/go-upgrade-rules"
+	"github.com/arangodb/kube-arangodb/pkg/apis/deployment"
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
 	"github.com/rs/zerolog"
@@ -85,18 +77,10 @@ func createRotateOrUpgradePlan(log zerolog.Logger, apiObject k8sutil.APIObject, 
 				newPlan = createUpgradeMemberPlan(log, m, group, "Version upgrade", spec.GetImage(), status,
 					!decision.AutoUpgradeNeeded)
 			} else {
-				// Upgrade is not needed, see if rotation is needed
-				if m.PodUID == "" {
-					rotNeeded, reason := podNeedsRotation(log, pod, apiObject, spec, group, status, m.ID, context)
-					if rotNeeded {
-						newPlan = createRotateMemberPlan(log, m, group, reason)
-					}
-				} else {
-					// Use new level of rotate logic
-					rotNeeded, reason := podNeedsRotationNew(log, pod, apiObject, spec, group, status, m, context)
-					if rotNeeded {
-						newPlan = createRotateMemberPlan(log, m, group, reason)
-					}
+				// Use new level of rotate logic
+				rotNeeded, reason := podNeedsRotation(log, pod, apiObject, spec, group, status, m, context)
+				if rotNeeded {
+					newPlan = createRotateMemberPlan(log, m, group, reason)
 				}
 			}
 
@@ -195,11 +179,11 @@ func podNeedsUpgrading(log zerolog.Logger, p core.Pod, spec api.DeploymentSpec, 
 	return upgradeDecision{UpgradeNeeded: false}
 }
 
-// podNeedsRotationNew returns true when the specification of the
+// podNeedsRotation returns true when the specification of the
 // given pod differs from what it should be according to the
 // given deployment spec.
 // When true is returned, a reason for the rotation is already returned.
-func podNeedsRotationNew(log zerolog.Logger, p core.Pod, apiObject metav1.Object, spec api.DeploymentSpec,
+func podNeedsRotation(log zerolog.Logger, p core.Pod, apiObject metav1.Object, spec api.DeploymentSpec,
 	group api.ServerGroup, status api.DeploymentStatus, m api.MemberStatus,
 	context PlanBuilderContext) (bool, string) {
 	if m.PodUID != p.UID {
@@ -235,257 +219,6 @@ func podNeedsRotationNew(log zerolog.Logger, p core.Pod, apiObject metav1.Object
 	return false, ""
 }
 
-// podNeedsRotation returns true when the specification of the
-// given pod differs from what it should be according to the
-// given deployment spec.
-// When true is returned, a reason for the rotation is already returned.
-func podNeedsRotation(log zerolog.Logger, p core.Pod, apiObject metav1.Object, spec api.DeploymentSpec,
-	group api.ServerGroup, status api.DeploymentStatus, id string,
-	context PlanBuilderContext) (bool, string) {
-	groupSpec := spec.GetServerGroupSpec(group)
-
-	// Check image pull policy
-	c, found := k8sutil.GetContainerByName(&p, k8sutil.ServerContainerName)
-	if found {
-		if c.ImagePullPolicy != spec.GetImagePullPolicy() {
-			return true, "Image pull policy changed"
-		}
-	} else {
-		return true, "Server container not found"
-	}
-
-	podImageInfo, found := status.Images.GetByImageID(c.Image)
-	if !found {
-		return false, "Server Image not found"
-	}
-
-	if group.IsExportMetrics() {
-		e, hasExporter := k8sutil.GetContainerByName(&p, k8sutil.ExporterContainerName)
-
-		if spec.Metrics.IsEnabled() {
-			if !hasExporter {
-				return true, "Exporter configuration changed"
-			}
-
-			if spec.Metrics.HasImage() {
-				if e.Image != spec.Metrics.GetImage() {
-					return true, "Exporter image changed"
-				}
-			}
-
-			if k8sutil.IsResourceRequirementsChanged(spec.Metrics.Resources, e.Resources) {
-				return true, "Resources requirements have been changed for exporter"
-			}
-		} else if hasExporter {
-			return true, "Exporter was disabled"
-		}
-	}
-
-	// Check arguments
-	expectedArgs := strings.Join(context.GetExpectedPodArguments(apiObject, spec, group, status.Members.Agents, id, podImageInfo.ArangoDBVersion), " ")
-	actualArgs := strings.Join(getContainerArgs(c), " ")
-	if expectedArgs != actualArgs {
-		log.Debug().
-			Str("actual-args", actualArgs).
-			Str("expected-args", expectedArgs).
-			Msg("Arguments changed. Rotation needed.")
-		return true, "Arguments changed"
-	}
-
-	// Check service account
-	if normalizeServiceAccountName(p.Spec.ServiceAccountName) != normalizeServiceAccountName(groupSpec.GetServiceAccountName()) {
-		return true, "ServiceAccountName changed"
-	}
-
-	// Check priorities
-	if groupSpec.PriorityClassName != p.Spec.PriorityClassName {
-		return true, "Pod priority changed"
-	}
-
-	// Check resource requirements
-	var resources core.ResourceRequirements
-	if groupSpec.HasVolumeClaimTemplate() {
-		resources = groupSpec.Resources // If there is a volume claim template compare all resources
-	} else {
-		resources = k8sutil.ExtractPodResourceRequirement(groupSpec.Resources)
-	}
-
-	if k8sutil.IsResourceRequirementsChanged(resources, k8sutil.GetArangoDBContainerFromPod(&p).Resources) {
-		return true, "Resource Requirements changed"
-	}
-
-	var memberStatus, _, _ = status.Members.MemberStatusByPodName(p.GetName())
-	if memberStatus.SideCarSpecs == nil {
-		memberStatus.SideCarSpecs = make(map[string]core.Container)
-	}
-
-	// Check for missing side cars in
-	for _, specSidecar := range groupSpec.GetSidecars() {
-		var stateSidecar core.Container
-		if stateSidecar, found = memberStatus.SideCarSpecs[specSidecar.Name]; !found {
-			return true, "Sidecar " + specSidecar.Name + " not found in running pod " + p.GetName()
-		}
-		if sideCarRequireRotation(specSidecar.DeepCopy(), &stateSidecar) {
-			return true, "Sidecar " + specSidecar.Name + " requires rotation"
-		}
-	}
-
-	for name := range memberStatus.SideCarSpecs {
-		var found = false
-		for _, specSidecar := range groupSpec.GetSidecars() {
-			if name == specSidecar.Name {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return true, "Sidecar " + name + " no longer in specification"
-		}
-	}
-
-	// Check for probe changes
-
-	// Readiness
-	if rotate, reason := compareProbes(pod.ReadinessSpec(group),
-		groupSpec.GetProbesSpec().GetReadinessProbeDisabled(),
-		groupSpec.GetProbesSpec().ReadinessProbeSpec,
-		c.ReadinessProbe); rotate {
-		return rotate, reason
-	}
-
-	// Liveness
-	if rotate, reason := compareProbes(pod.LivenessSpec(group),
-		groupSpec.GetProbesSpec().LivenessProbeDisabled,
-		groupSpec.GetProbesSpec().LivenessProbeSpec,
-		c.LivenessProbe); rotate {
-		return rotate, reason
-	}
-
-	// Volumes
-	if rotate, reason := compareVolumes(groupSpec, p); rotate {
-		return rotate, reason
-	}
-
-	if rotate, reason := compareVolumeMounts(groupSpec, c); rotate {
-		return rotate, reason
-	}
-
-	return false, ""
-}
-
-func compareProbes(probe pod.Probe, groupProbeDisabled *bool, groupProbeSpec *api.ServerGroupProbeSpec, containerProbe *core.Probe) (bool, string) {
-	if !probe.CanBeEnabled {
-		if containerProbe != nil {
-			return true, "Probe needs to be disabled"
-		}
-		// Nothing to do - we cannot enable probe and probe is disabled
-		return false, ""
-	}
-
-	enabled := probe.EnabledByDefault
-
-	if groupProbeDisabled != nil {
-		enabled = !*groupProbeDisabled
-	}
-
-	if enabled && containerProbe == nil {
-		// We expected probe to be enabled but it is disabled in container
-		return true, "Enabling probe"
-	}
-
-	if !enabled {
-		if containerProbe != nil {
-			// We expected probe to be disabled but it is enabled in container
-			return true, "Disabling probe"
-		}
-
-		// Nothing to do - probe is disabled and it should stay as it is
-		return false, ""
-	}
-
-	if groupProbeSpec.GetTimeoutSeconds(containerProbe.TimeoutSeconds) != containerProbe.TimeoutSeconds {
-		return true, "Timeout seconds does not match"
-	}
-
-	if groupProbeSpec.GetPeriodSeconds(containerProbe.PeriodSeconds) != containerProbe.PeriodSeconds {
-		return true, "Period seconds does not match"
-	}
-
-	// Recreate probe if timeout seconds are different
-	if groupProbeSpec.GetInitialDelaySeconds(containerProbe.InitialDelaySeconds) != containerProbe.InitialDelaySeconds {
-		return true, "Initial delay seconds does not match"
-	}
-
-	// Recreate probe if timeout seconds are different
-	if groupProbeSpec.GetFailureThreshold(containerProbe.FailureThreshold) != containerProbe.FailureThreshold {
-		return true, "Failure threshold does not match"
-	}
-
-	// Recreate probe if timeout seconds are different
-	if groupProbeSpec.GetSuccessThreshold(containerProbe.SuccessThreshold) != containerProbe.SuccessThreshold {
-		return true, "Success threshold does not match"
-	}
-
-	return false, ""
-}
-
-func compareVolumes(spec api.ServerGroupSpec, pod core.Pod) (bool, string) {
-	if len(spec.Volumes) == 0 {
-		return false, ""
-	}
-
-	currentVolumes := map[string]core.Volume{}
-
-	for _, volume := range pod.Spec.Volumes {
-		currentVolumes[volume.Name] = volume
-	}
-
-	for _, expectedVolumeTemplate := range spec.Volumes {
-		expectedVolume := expectedVolumeTemplate.Volume()
-
-		currentVolume, ok := currentVolumes[expectedVolume.Name]
-		if !ok {
-			return true, fmt.Sprintf("Volume %s is not mount. Rotating", expectedVolume.Name)
-		}
-
-		if !equality.Semantic.DeepDerivative(expectedVolume, currentVolume) {
-			return true, fmt.Sprintf("Volume %s needs to be updated", expectedVolume.Name)
-		}
-	}
-
-	return false, ""
-}
-
-func compareVolumeMounts(spec api.ServerGroupSpec, container core.Container) (bool, string) {
-	if len(container.VolumeMounts) < len(spec.VolumeMounts) {
-		return true, "Missing volume mounts in container"
-	}
-
-	// Get compared mounts
-	mounts := container.VolumeMounts
-	if len(mounts) > 0 {
-		if container.VolumeMounts[len(mounts)-1].MountPath == "/var/run/secrets/kubernetes.io/serviceaccount" {
-			// Remove last mount added by ServiceAccount from compare
-			mounts = mounts[0 : len(mounts)-1]
-		}
-	}
-
-	if len(spec.VolumeMounts) > len(mounts) {
-		return true, "Missing volume mounts in container"
-	}
-
-	mounts = mounts[len(mounts)-len(spec.VolumeMounts):]
-
-	// Now we can compare lists
-	for id, mount := range mounts {
-		if !equality.Semantic.DeepDerivative(spec.VolumeMounts[id].VolumeMount(), mount) {
-			return true, fmt.Sprintf("Mount with if %d does not match - got %s, expected %s", id, mount.Name, spec.VolumeMounts[id].Name)
-		}
-	}
-
-	return false, ""
-}
-
 // clusterReadyForUpgrade returns true if the cluster is ready for the next update, that is:
 // 	- all shards are in sync
 // 	- all members are ready and fine
@@ -493,19 +226,6 @@ func clusterReadyForUpgrade(context PlanBuilderContext) bool {
 	status, _ := context.GetStatus()
 	allInSync := context.GetShardSyncStatus()
 	return allInSync && status.Conditions.IsTrue(api.ConditionTypeReady)
-}
-
-// sideCarRequireRotation checks if side car requires rotation including default parameters
-func sideCarRequireRotation(wanted, given *core.Container) bool {
-	return !reflect.DeepEqual(wanted, given)
-}
-
-// normalizeServiceAccountName replaces default with empty string, otherwise returns the input.
-func normalizeServiceAccountName(name string) string {
-	if name == "default" {
-		return ""
-	}
-	return ""
 }
 
 // createUpgradeMemberPlan creates a plan to upgrade (stop-recreateWithAutoUpgrade-stop-start) an existing
@@ -532,11 +252,4 @@ func createUpgradeMemberPlan(log zerolog.Logger, member api.MemberStatus,
 		}, plan...)
 	}
 	return plan
-}
-
-func getContainerArgs(c core.Container) []string {
-	if len(c.Command) >= 1 {
-		return c.Command[1:]
-	}
-	return c.Args
 }